Trait reth_trie::hashed_cursor::HashedCursorFactory

pub trait HashedCursorFactory {
    type AccountCursor: HashedAccountCursor;
    type StorageCursor: HashedStorageCursor;

    // Required methods
    fn hashed_account_cursor(
        &self
    ) -> Result<Self::AccountCursor, DatabaseError>;
    fn hashed_storage_cursor(
        &self
    ) -> Result<Self::StorageCursor, DatabaseError>;
}

The factory trait for creating cursors over the hashed state.

Required Associated Types

type AccountCursor: HashedAccountCursor

The hashed account cursor type.

type StorageCursor: HashedStorageCursor

The hashed storage cursor type.

Required Methods

fn hashed_account_cursor(&self) -> Result<Self::AccountCursor, DatabaseError>

Returns a cursor for iterating over all hashed accounts in the state.

fn hashed_storage_cursor(&self) -> Result<Self::StorageCursor, DatabaseError>

Returns a cursor for iterating over all hashed storage entries in the state.