Trait reth_trie::hashed_cursor::HashedStorageCursor

pub trait HashedStorageCursor {
    // Required methods
    fn is_storage_empty(&mut self, key: B256) -> Result<bool, DatabaseError>;
    fn seek(
        &mut self,
        key: B256,
        subkey: B256
    ) -> Result<Option<StorageEntry>, DatabaseError>;
    fn next(&mut self) -> Result<Option<StorageEntry>, DatabaseError>;
}

The cursor for iterating over hashed storage entries.

Required Methods

fn is_storage_empty(&mut self, key: B256) -> Result<bool, DatabaseError>

Returns true if there are no entries for a given key.

fn seek( &mut self, key: B256, subkey: B256 ) -> Result<Option<StorageEntry>, DatabaseError>

Seek an entry greater or equal to the given key/subkey and position the cursor there.

fn next(&mut self) -> Result<Option<StorageEntry>, DatabaseError>

Move the cursor to the next entry and return it.