Trait reth_trie::trie_cursor::TrieCursor

pub trait TrieCursor {
    type Key: From<Vec<u8>>;

    // Required methods
    fn seek_exact(
        &mut self,
        key: Self::Key
    ) -> Result<Option<(Vec<u8>, BranchNodeCompact)>, DatabaseError>;
    fn seek(
        &mut self,
        key: Self::Key
    ) -> Result<Option<(Vec<u8>, BranchNodeCompact)>, DatabaseError>;
    fn current(&mut self) -> Result<Option<TrieKey>, DatabaseError>;
}

A cursor for navigating a trie that works with both Tables and DupSort tables.

Required Associated Types

type Key: From<Vec<u8>>

The key type of the cursor.

Required Methods

fn seek_exact( &mut self, key: Self::Key ) -> Result<Option<(Vec<u8>, BranchNodeCompact)>, DatabaseError>

Move the cursor to the key and return if it is an exact match.

fn seek( &mut self, key: Self::Key ) -> Result<Option<(Vec<u8>, BranchNodeCompact)>, DatabaseError>

Move the cursor to the key and return a value matching of greater than the key.

fn current(&mut self) -> Result<Option<TrieKey>, DatabaseError>

Get the current entry.